chore: abtract social login flows to SocialUtil
