+++ packages/controllers/src/utils/TypeUtil.ts
@@ -89,7 +89,6 @@ export type SocialProvider =
   | 'x'
   | 'discord'
   | 'farcaster'
-
 export type Connector = {
   id: string
   type: ConnectorType
